referrer-policy/generic/subresource-test/attr-referrer-invalid-value.html

<!DOCTYPE html>
<html>
  <head>
    <title>Invalid referrerpolicy attribute value</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <h1>Invalid referrerpolicy attribute value</h1>
    <pre>Running...</pre>

    <script>
      var test = test("Invalid referrerpolicy values not reflected");
      var elements = ["iframe", "img", "a", "area"];
      for (var i = 0; i < elements.length; i++) {
          var elem = document.createElement(elements[i]);
          elem.referrerpolicy = "unsafe-url";
          assert_equals("unsafe-url", elem.referrerpolicy);
          elem.referrerpolicy = "not-valid-value";
          assert_equals("", elem.referrerpolicy);
      }
    </script>
  </body>
</html>
