LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/subresource-test/attr-referrer-invalid-value-expected.txt

Invalid referrerPolicy attribute value

Running...

FAIL Invalid referrerpolicy values not reflected assert_equals: expected "" but got "not-valid-value"
